Add amenity placement assistant for automated bench/bin placement
Implements intelligent amenity placement that:
- Detects nauseous ride exits and prioritizes benches nearby
- Detects food shops and places bins for litter coverage
- Uses path connectivity (matching game's algorithm) for accurate exit detection
- Applies spreading penalty to distribute amenities evenly
- Respects placement constraints (slopes, junctions, edges, ownership)
- Penalizes bins near ride exits without food (bins don't help nausea)

Commands: amenities analyze, amenities auto --budget <amount>

+++ src/openrct2/drawing/Drawing.cpp
@@ -62,7 +62,7 @@ PaletteIndex PaletteMap::Blend(PaletteIndex src, PaletteIndex dst) const
 {
 #ifdef _DEBUG
     // src = 0 would be transparent so there is no blend palette for that, hence (src - 1)
-    assert(src != PaletteIndex::pi0 && (EnumValue(src) - 1) < _numMaps);
+    assert(src != PaletteIndex::pi0 && static_cast<size_t>(EnumValue(src) - 1) < _numMaps);
     assert(EnumValue(dst) < _mapLength);
 #endif
     auto idx = ((EnumValue(src) - 1) * 256) + EnumValue(dst);
